enh: use REGISTERED_CUSTOM_CHECKS for attribute lookup, add dir method
