+++ tests/core/checks_fixtures.py
@@ -1,0 +1,32 @@
+"""Pytest fixtures for testing custom checks."""
+import unittest.mock as mock
+import pytest
+import pandas as pd
+
+import pandera as pa
+import pandera.extensions as pa_ext
+
+__all__ = "custom_check_teardown", "extra_registered_checks"
+
+
+@pytest.fixture(scope="function")
+def custom_check_teardown():
+    """Remove all custom checks after execution of each pytest function."""
+    yield
+    for check_name in list(pa.Check.REGISTERED_CUSTOM_CHECKS):
+        del pa.Check.REGISTERED_CUSTOM_CHECKS[check_name]
+
+
+@pytest.fixture(scope="function")
+def extra_registered_checks():
+    """temporarily registers custom checks onto the Check class"""
+    # pylint: disable=unused-variable
+    with mock.patch(
+        "pandera.Check.REGISTERED_CUSTOM_CHECKS", new_callable=dict
+    ):
+        # register custom checks here
+        @pa_ext.register_check_method()
+        def no_param_check(_: pd.DataFrame) -> bool:
+            return True
+
+        yield
