+++ tests/core/conftest.py
@@ -1,0 +1,4 @@
+"""Registers fixtures for core"""
+
+# pylint: disable=unused-import
+from .checks_fixtures import custom_check_teardown, extra_registered_checks
